refactored logger part 3
